Separate header and footer content from index file
Moved header and footer content from index file to their respective
PHP files. Used php tag and getheader and getfooter to recall them
and combine them. The DOCTYPE and head tags are in the header file,
as is the open tag of the container <div>. The closing tag of the
container <div> is in the footer file.