Modify files for Data Context/Add Sidebar
Addition to index.php file: Added the php that recalls the individual
page titles and a paragraph excerpt for homepage.
Addition to page.php: created all content in file relating to Data
Context Loop. Also added the php code that recalls the sidebar (see
comments in code). Changed the main content class to “nine column” and
the sidebar content is “three columns”. This changes the grid system to
include both (for a total of 12 columns).
Addition to sidebar.php: added content that will display in sidebar on
homepage.
Addition to single.php: added content that will display in single pages.
This is basically the same content that is on page.php.